+++ packages/http/src/Middleware.ts
@@ -1,5 +1,6 @@
 import { HttpContext } from './Contracts/HttpContract';
+import { IMiddleware } from '@h3ravel/shared';
 
-export abstract class Middleware {
+export abstract class Middleware implements IMiddleware {
     abstract handle (context: HttpContext, next: () => Promise<unknown>): Promise<unknown>
 }