+++ packages/http/src/Request.ts
@@ -1,7 +1,8 @@
 import { getQuery, getRouterParams, readBody, type H3Event } from 'h3'
 import { DotNestedKeys, DotNestedValue, safeDot } from '@h3ravel/support'
+import { IRequest } from '@h3ravel/shared'
 
-export class Request {
+export class Request implements IRequest {
     private readonly event: H3Event
 
     constructor(event: H3Event) {