+++ packages/shared/src/Contracts/IServiceProvider.ts
@@ -1,0 +1,13 @@
+export interface IServiceProvider {
+    /**
+     * Register bindings to the container.
+     * Runs before boot().
+     */
+    register (): void | Promise<void>
+
+    /**
+     * Perform post-registration booting of services.
+     * Runs after all providers have been registered.
+     */
+    boot?(): void | Promise<void>
+}